refactor: 1. merge connection_jitter 2. merge sub
