+++ src/cmd/src/cli-command/mqtt/admin.rs
@@ -19,7 +19,6 @@ use clap::builder::{
 };
 use clap::{arg, Parser};
 use cli_command::mqtt::MqttActionType;
-use common_base::enum_type::common_enum::SortType;
 use common_base::enum_type::sort_type::SortType;
 use protocol::broker_mqtt::broker_mqtt_admin::{CreateUserRequest, DeleteUserRequest};
 use protocol::broker_mqtt::broker_mqtt_admin::{
